+++ CHANGELOG.md
@@ -4,6 +4,10 @@ Borcama'nın kullanıcıya, yönetime ve entegrasyonlara yansıyan değişiklikl
 
 ## Unreleased
 
+### Düzeltildi
+
+- Google Ads kayıt dönüşümü, doğrulanmamış oturumlar yerine yalnız e-posta doğrulamasını tamamlayan uygulama kaynaklı hesaplarda gönderilir.
+
 ## [1.42.4] - 2026-09-04
 
 ### Düzeltildi